resolved most of the warnings
Updated the app to use the navigation bar to move to different apps.

Created the OnSwipeListener class. Allows for swipe to be used.

Updated the members side bar to use Swipes

Added the add member, settings, add team, qnd color picker menus.

Added the ActivityVariables class to hold the context, View and Activity for each fragment or activity.